pyproject.toml
[project]
name = "livekit-poc"
version = "0.1.0"
description = "livekit-poc"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "livekit-agents[google,deepgram,openai,silero,turn-detector,elevenlabs]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
